+++ api-server/scanner-lib/src/sync/local_state.rs
